move i18n_block to its own file
